Fixed serialization for UINT_8, INT_8, UINT_16 and INT16 to be based on INT32

+++ src/Parquet/Data/Concrete/ByteDataTypeHandler.cs
@@ -16,12 +16,12 @@ namespace Parquet.Data.Concrete
 
       protected override byte ReadSingle(BinaryReader reader, Thrift.SchemaElement tse, int length)
       {
-         return reader.ReadByte();
+         return (byte) reader.ReadInt32();
       }
 
       protected override void WriteOne(BinaryWriter writer, byte value)
       {
-         writer.Write(value);
+         writer.Write((int) value);
       }
    }
 }
